refactor: move backup module and input module to coreutils_core
