+++ mv/src/main.rs
@@ -1,132 +1,9 @@
-use self::{backup::*, input::Input};
+use coreutils_core::{backup::*, input::*};
 use clap::{load_yaml, App, ArgMatches};
 use std::{
     fs,
     path::{Path, PathBuf},
 };
-
-// TODO(gab): Extract this to core because cp, ln, etc use backups
-pub mod backup {
-    use regex::Regex;
-    use std::{
-        fs,
-        io::{Error, ErrorKind},
-        path::PathBuf,
-    };
-
-    #[derive(Debug, Clone, PartialEq)]
-    pub enum BackupMode {
-        None,
-        Numbered,
-        Existing,
-        Simple,
-    }
-
-    impl BackupMode {
-        pub fn from_string(string: &str) -> BackupMode {
-            match string {
-                "none" | "off" => BackupMode::None,
-                "numbered" | "t" => BackupMode::Numbered,
-                "existing" | "nil" => BackupMode::Existing,
-                "simple" | "never" => BackupMode::Simple,
-                _ => BackupMode::Existing,
-            }
-        }
-    }
-
-    pub fn create_numbered_backup(file: &PathBuf) -> Result<PathBuf, Error> {
-        let mut index = 1_u64;
-        loop {
-            if index == std::u64::MAX {
-                return Err(Error::new(
-                    ErrorKind::AlreadyExists,
-                    "Cannot create backup: too many backup files",
-                ));
-            }
-
-            let new = file.with_extension(format!("~{}~", index));
-            if !new.exists() {
-                match fs::rename(file, &new) {
-                    Ok(()) => return Ok(new),
-                    Err(err) => return Err(err),
-                };
-            }
-
-            index += 1;
-        }
-    }
-
-    pub fn create_existing_backup(file: &PathBuf, suffix: &String) -> Result<PathBuf, Error> {
-        let mut has_numbered_backup = false;
-        let regex = Regex::new(r"~\d+~").unwrap();
-        let parent = file.parent().unwrap();
-        for entry in parent.read_dir().unwrap() {
-            if let Ok(entry) = entry {
-                if let Some(ext) = entry.path().extension() {
-                    if regex.is_match(ext.to_str().unwrap()) {
-                        has_numbered_backup = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if has_numbered_backup {
-            return create_numbered_backup(file);
-        } else {
-            return create_simple_backup(file, suffix);
-        }
-    }
-
-    pub fn create_simple_backup(file: &PathBuf, suffix: &String) -> Result<PathBuf, Error> {
-        let new = PathBuf::from(format!("{}{}", file.display(), suffix));
-
-        match fs::rename(file, &new) {
-            Ok(()) => Ok(new),
-            Err(error) => Err(error),
-        }
-    }
-}
-
-// TODO(gab): extract to core because a tonne of core utils use this
-mod input {
-    use std::{io, io::prelude::*, process};
-
-    #[derive(Debug)]
-    pub struct Input(String);
-
-    impl Input {
-        pub fn new() -> Self {
-            let mut line = String::new();
-            match io::stdin().lock().read_line(&mut line) {
-                Ok(_) => {},
-                Err(err) => {
-                    eprintln!("rm: cannot read input: {}", err);
-                    process::exit(1);
-                },
-            };
-
-            Input(line)
-        }
-
-        pub fn with_msg(msg: &str) -> Self {
-            print!("{}", msg);
-
-            if let Err(err) = io::stdout().lock().flush() {
-                eprintln!("rm: could not flush stdout: {}", err);
-                process::exit(1);
-            }
-
-            Self::new()
-        }
-
-        pub fn is_affirmative(&self) -> bool {
-            let input = self.0.trim().to_uppercase();
-
-            input == "Y" || input == "YES" || input == "1"
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 enum OverwriteMode {
